requirements.txt
python-dotenv
streamlit
PyPDF2
hdbcli
faiss-cpu
generative-ai-hub-sdk[all]